# api-docs.yml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Video CRUD 
# during dev, should point to your local machine
host: localhost:8081
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /video:    
    post:
      description: Create video object
      parameters:
        - name: video
          description: Video properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/VideoObject"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/VideoObjectResponse"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"       
    get:
      description: Get the video object list     
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/VideoListResponse"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
  /video/{id}:      
    get:
      description: Get the video object by id
      parameters:
        - name: id
          type: string
          in: path
          required: true     
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/VideoObjectResponse"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Update video object by id
      parameters:
        - name: id
          type: string
          in: path
          required: true
        - name: video
          description: Video properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/VideoObject"       
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralSuccessResponse"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Delete video object by id
      parameters:
        - name: id
          type: string
          in: path
          required: true     
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralSuccessResponse"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"                                            
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  ErrorResponse:
    required:
      - message
      - status 
    properties:
      status:
        type: string
      message:
        type: string
  GeneralSuccessResponse:
    required:
      - message
      - status 
    properties:
      status:
        type: string
      message:
        type: string      
  VideoObjectResponse:
    required: 
      - _id     
      - name 
      - keywords
      - createdDate 
    properties:     
      _id:
        type: string
      name:
        type: string
      keywords:
        type: array 
        items: 
          type: string
      createdDate:
        type: string      
  VideoListResponse:
    required:
      - status
      - message 
      - data
    properties:
      status:
        type: string
      message:
        type: string
      # The array of video obj
      data:
        type: array
        items: 
          type: object
          properties:
            _id:
              type: string
            name:
              type: string
            keywords:
              type: array 
              items: 
                type: string
            createdDate:
              type: string  
  VideoObject:
    required:     
      - name 
      - keywords
    properties:     
      name:
        type: string
      keywords:
        type: array 
        items: 
          type: string
      
                 